+++ expensecalendar-backend/src/main/java/com/ozeken/expensecalendar/mapper/ExpenseMapper.java
@@ -1,4 +1,4 @@
-package com.ozeken.expensecalendar.repository;
+package com.ozeken.expensecalendar.mapper;
 
 import java.util.List;
