+++ expensecalendar-backend/src/main/java/com/ozeken/expensecalendar/mapper/UserMapper.java
@@ -1,4 +1,4 @@
-package com.ozeken.expensecalendar.repository;
+package com.ozeken.expensecalendar.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
